Changed delimiters for asciimath from \$ to <ascii> ... </ascii>
