+++ lib/markdown/md-ascii/index.d.ts
@@ -1,0 +1,3 @@
+export declare const asciiMath: (state: any, silent: any) => boolean;
+export declare const backtickAsAsciiMath: (state: any, silent: any) => boolean;
+export declare const renderAsciiMath: (tokens: any, idx: any, options: any) => string;
