+++ lib/markdown/mdPluginRaw.js
@@ -4,6 +4,7 @@ exports.findEndMarkerPos = void 0;
 var mathjax_1 = require("../mathjax/");
 var tabular_1 = require("./md-inline-rule/tabular");
 var render_tabular_1 = require("./md-renderer-rules/render-tabular");
+var md_ascii_1 = require("./md-ascii");
 var utils_1 = require("./utils");
 var mathNumber = [];
 var attr_name = '[a-zA-Z_:][a-zA-Z0-9:._-]*';
@@ -292,83 +293,6 @@ function refs(state, silent) {
     state.pos = nextPos;
     return true;
 }
-var asciiMath = function (state, silent) {
-    var notRenderAsciiMath = state.md.options.mathJax
-        && Object(state.md.options.mathJax).hasOwnProperty('asciiMath')
-        && state.md.options.mathJax.asciiMath === false;
-    if (notRenderAsciiMath) {
-        return false;
-    }
-    var pos, afterStartMarker, startMathPos = state.pos;
-    var asciiMarker = "\\$";
-    if (state.src.charCodeAt(startMathPos) !== 0x5c /* \ */) {
-        return false;
-    }
-    pos = ++startMathPos;
-    afterStartMarker = state.src.charCodeAt(pos);
-    if (afterStartMarker !== 0x24 /* $ */) {
-        return false;
-    }
-    if (state.src.charCodeAt(++startMathPos) === 0x24 /* $ */) {
-        return false;
-    }
-    var endMarkerPos = exports.findEndMarkerPos(state.src, asciiMarker, startMathPos);
-    if (endMarkerPos === -1) {
-        return false;
-    }
-    if (state.src.charCodeAt(endMarkerPos - 1) === 0x5c /* \ */) {
-        return false;
-    }
-    var nextPos = endMarkerPos + asciiMarker.length;
-    if (!silent) {
-        var token = state.push("ascii_math", 0);
-        token.content = state.src.slice(startMathPos, endMarkerPos);
-        if (state.env.tabulare) {
-            token.return_asciimath = true;
-        }
-        if (state.md.options.forLatex) {
-            token.markup = startMathPos;
-        }
-    }
-    state.pos = nextPos;
-    return true;
-};
-var backtickAsAsciiMath = function (state, silent) {
-    var useBacktick = state.md.options.mathJax && state.md.options.mathJax.asciiMath && state.md.options.mathJax.asciiMath.useBacktick;
-    if (!useBacktick) {
-        return false;
-    }
-    var start, marker, matchStart, matchEnd, token, pos = state.pos, ch = state.src.charCodeAt(pos);
-    if (ch !== 0x60 /* ` */) {
-        return false;
-    }
-    start = pos;
-    pos++;
-    if (state.src.charCodeAt(pos) === 0x60 /* ` */) {
-        return false;
-    }
-    marker = state.src.slice(start, pos);
-    matchStart = matchEnd = pos;
-    while ((matchStart = state.src.indexOf('`', matchEnd)) !== -1) {
-        matchEnd = matchStart + 1;
-        if (matchEnd - matchStart === marker.length) {
-            if (!silent) {
-                token = state.push('ascii_math', 0);
-                token.markup = marker;
-                token.content = state.src.slice(pos, matchStart)
-                    .replace(/[ \n]+/g, ' ')
-                    .trim();
-            }
-            state.pos = matchEnd;
-            return true;
-        }
-    }
-    if (!silent) {
-        state.pending += marker;
-    }
-    state.pos += marker.length;
-    return true;
-};
 function simpleMath(state, silent) {
     var pos, afterStartMarker, startMathPos = state.pos;
     var endMarker = "$";
@@ -510,49 +434,6 @@ var checkReference = function (data) {
         math: data.replace(/\\label\{([^}]*)\}/, "")
     };
 };
-var getWidthFromDocument = function (cwidth) {
-    if (cwidth === void 0) { cwidth = 1200; }
-    try {
-        var el_container = document.getElementById('container-ruller');
-        return el_container ? el_container.offsetWidth : 1200;
-    }
-    catch (e) {
-        return cwidth;
-    }
-};
-var renderAsciiMath = function (tokens, idx, options) {
-    var token = tokens[idx];
-    var mathEquation = null;
-    var math = token.content;
-    try {
-        var cwidth = 1200;
-        if (options && options.width && options.width > 0) {
-            cwidth = options.width;
-        }
-        else {
-            cwidth = getWidthFromDocument(cwidth);
-        }
-        mathEquation = mathjax_1.MathJax.AsciiMathToSvg(math, { display: false, metric: { cwidth: cwidth },
-            outMath: options.outMath, mathJax: options.mathJax, forDocx: options.forDocx });
-    }
-    catch (e) {
-        console.error('ERROR MathJax =>', e.message, e);
-        if (options.outMath && options.outMath.not_catch_errors) {
-            throw ({
-                message: e.message,
-                error: e
-            });
-        }
-        if (token.type === 'display_mathML' || token.type === 'inline_mathML') {
-            mathEquation = "<span class=\"math-error\">" + math + "</span>";
-        }
-        else {
-            mathEquation = math;
-            return "<p class=\"math-error\">" + mathEquation + "</p>";
-        }
-    }
-    return "<span class=\"math-inline ascii\">" + mathEquation + "</span>";
-};
 var renderMath = function (a, token, options) {
     var _a = checkReference(token.content), tagId = _a.tagId, math = _a.math;
     var isBlock = token.type !== 'inline_math';
@@ -564,7 +445,7 @@ var renderMath = function (a, token, options) {
             cwidth = options.width;
         }
         else {
-            cwidth = getWidthFromDocument(cwidth);
+            cwidth = utils_1.getWidthFromDocument(cwidth);
         }
         if (token.type === 'display_mathML' || token.type === 'inline_mathML') {
             mathEquation = mathjax_1.MathJax.TypesetMathML(math);
@@ -932,8 +813,8 @@ exports.default = (function (options) {
         md.inline.ruler.before("escape", "multiMath", multiMath);
         md.inline.ruler.before("multiMath", "inlineTabular", tabular_1.inlineTabular);
         md.inline.ruler.push("simpleMath", simpleMath);
-        md.inline.ruler.before("multiMath", "asciiMath", asciiMath);
-        md.inline.ruler.before("asciiMath", "backtickAsAsciiMath", backtickAsAsciiMath);
+        md.inline.ruler.before("multiMath", "asciiMath", md_ascii_1.asciiMath);
+        md.inline.ruler.before("asciiMath", "backtickAsAsciiMath", md_ascii_1.backtickAsAsciiMath);
         Object.keys(mapping).forEach(function (key) {
             md.renderer.rules[key] = function (tokens, idx, options, env, slf) {
                 switch (tokens[idx].type) {
@@ -950,7 +831,7 @@ exports.default = (function (options) {
                     case "usepackage_geometry":
                         return renderUsepackage(tokens[idx], options);
                     case "ascii_math":
-                        return renderAsciiMath(tokens, idx, options);
+                        return md_ascii_1.renderAsciiMath(tokens, idx, options);
                     default:
                         return renderMath(tokens, tokens[idx], options);
                 }
