+++ lib/markdown/utils.d.ts
@@ -1,5 +1,6 @@
 export declare const endTag: (arg: string) => RegExp;
 export declare const getTextWidth: () => number;
+export declare const getWidthFromDocument: (cwidth?: number) => number;
 export declare const isNotBackticked: (str: string, tag: string) => boolean;
 export declare const includesSimpleMathTag: (str: string, tag?: string) => boolean;
 export declare const includesMultiMathBeginTag: (str: any, tag: any) => RegExp | null;
