+++ lib/markdown/utils.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.uid = exports.arrayResortFromElement = exports.arrayDelElement = exports.arraysCompare = exports.includesMultiMathTag = exports.includesMultiMathBeginTag = exports.includesSimpleMathTag = exports.isNotBackticked = exports.getTextWidth = exports.endTag = void 0;
+exports.uid = exports.arrayResortFromElement = exports.arrayDelElement = exports.arraysCompare = exports.includesMultiMathTag = exports.includesMultiMathBeginTag = exports.includesSimpleMathTag = exports.isNotBackticked = exports.getWidthFromDocument = exports.getTextWidth = exports.endTag = void 0;
 var tslib_1 = require("tslib");
 exports.endTag = function (arg) {
     return new RegExp('\\\\end\s{0,}\{(' + arg + ')\}');
@@ -8,6 +8,16 @@ exports.endTag = function (arg) {
 exports.getTextWidth = function () {
     var el_container = document ? document.getElementById('container-ruller') : null;
     return el_container ? el_container.offsetWidth : 800;
+};
+exports.getWidthFromDocument = function (cwidth) {
+    if (cwidth === void 0) { cwidth = 1200; }
+    try {
+        var el_container = document.getElementById('container-ruller');
+        return el_container ? el_container.offsetWidth : 1200;
+    }
+    catch (e) {
+        return cwidth;
+    }
 };
 exports.isNotBackticked = function (str, tag) {
     var pos = 0, max = str.length, ch, escapes = 0, backTicked = false, lastBackTick = 0;
