+++ lib/mathjax/mathJaxConfig.js
@@ -24,7 +24,7 @@ var MathJaxConfig = {
         processRefs: true,
     },
     asciimath: {
-        delimiters: [['\\$', '\\$'], ['`', '`']]
+        delimiters: [['<ascii>', '</ascii>'], ['`', '`']]
     },
     MathML: {
         parseAs: "html",
